+++ frontend/src/components/__tests__/FilterBar.test.tsx
@@ -1,7 +1,7 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest';
+import { describe, it, expect, vi } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { FilterBar } from '../frontend/src/components/FilterBar';
+import { FilterBar } from '../FilterBar';
 
 describe('FilterBar Component', () => {
   const mockOnFilterChange = vi.fn();
