+++ frontend/src/components/__tests__/api.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
-import { api } from '../api';
+import { api } from '../../utils/api';
 
 // Mock fetch for testing
 global.fetch = vi.fn();
